doc: Explain how to change browser for test
It's possible to use many browser for the tests.
As angular2 needs phantomjs2 instead of phantomjs, it's important to let users know
about it.
Tests are running fine in chrome, that's why it's the default browser for now.
It might change as soon as phantomjs2 may be installed in an easier way.